+++ core/stats/calculationStats.h
@@ -178,7 +178,6 @@ public:
     virtual void addSingleStat(string name, const SingleStat &stat)
     {
         UnitedStat unitedStat;
-        //typename map<string, UnitedStat>::iterator uit;
         map<string, UnitedStat>::iterator uit;
         uit = sumValues.find(name);
         if (uit != sumValues.end())
@@ -205,7 +204,6 @@ public:
 
     virtual int maximumCaptionLength() const
     {
-        //typename StatsMap::const_iterator uit;
         StatsMap::const_iterator uit;
         int maxCaptionLen = 0;
         for (uit = this->sumValues.begin(); uit != this->sumValues.end(); ++uit )
@@ -228,7 +226,7 @@ template <class StreamType>
         int maxCaptionLen = maximumCaptionLength();
 
         /* Flags */
-        int size = this->sumValues.size();
+        int size = (int)this->sumValues.size();
         bool *isPrinted = new bool[size];
         for (int i = 0; i < size; i++)
         {
